Add download and delete functionality

+++ File.cs
@@ -1,12 +1,10 @@
-using System;
-
 namespace EventsManagement.Core
 {
   public class EventFile
   {
     public int Id { get; set; }
     public string Name { get; set; }
-    public int Size { get; set; }
+    public long Size { get; set; }
     public string Path { get; set; }
     public string Type { get; set; }
